eval/pyproject.toml
[project]
name = "researchpipe-eval"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "markdown>=3.10.2",
    "openai>=2.33.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.24",
    "tenacity>=9.1.4",
]
